_legacy_MVP1/refactored_pipelines/requirements.txt
# Core Libraries
pandas
numpy
scikit-learn

# API Clients
openai
google-generativeai
supabase
psycopg2-binary

# Other utilities
requests
beautifulsoup4
feedparser
tqdm
python-dotenv
